requirements.txt
anthropic>=0.39.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
